New namespace for SynchronizationObject.
