+++ lib/concurrent.rb
@@ -1,6 +1,6 @@
 require 'concurrent/version'
 
-require 'concurrent/synchronization'
+require 'concurrent/synchronization_object'
 require 'concurrent/at_exit'
 
 require 'concurrent/configuration'
