+++ lib/concurrent/at_exit.rb
@@ -1,11 +1,11 @@
 require 'concurrent/logging'
-require 'concurrent/synchronization'
+require 'concurrent/synchronization_object'
 
 module Concurrent
 
   # Provides ability to add and remove handlers to be run at `Kernel#at_exit`, order is undefined.
   # Each handler is executed at most once.
-  class AtExitImplementation < Synchronization::Object
+  class AtExitImplementation < SynchronizationObject
     include Logging
 
     # Add a handler to be run at `Kernel#at_exit`
