+++ lib/concurrent/executor/safe_task_executor.rb
@@ -1,4 +1,4 @@
-require 'concurrent/synchronization'
+require 'concurrent/synchronization_object'
 
 module Concurrent
 
@@ -6,7 +6,7 @@ module Concurrent
   # success - indicating if the callable has been executed without errors
   # value - filled by the callable result if it has been executed without errors, nil otherwise
   # reason - the error risen by the callable if it has been executed with errors, nil otherwise
-  class SafeTaskExecutor < Synchronization::Object
+  class SafeTaskExecutor < SynchronizationObject
 
     def initialize(task, opts = {})
       super()
