+++ lib/concurrent/struct/mutable_struct.rb
@@ -1,5 +1,5 @@
 require 'concurrent/struct/abstract_struct'
-require 'concurrent/synchronization'
+require 'concurrent/synchronization_object'
 
 module Concurrent
 
@@ -206,10 +206,10 @@ module Concurrent
       FACTORY.define_struct(clazz_name, args, &block)
     end
 
-    FACTORY = Class.new(Synchronization::Object) do
+    FACTORY = Class.new(SynchronizationObject) do
       def define_struct(name, members, &block)
         synchronize do
-          clazz = AbstractStruct.define_struct_class(MutableStruct, Synchronization::Object, name, members, &block)
+          clazz = AbstractStruct.define_struct_class(MutableStruct, SynchronizationObject, name, members, &block)
           members.each_with_index do |member, index|
             clazz.send(:define_method, member) do
               synchronize { @values[index] }
