--- lib/concurrent/synchronization.rb
@@ -1,14 +1,0 @@
-require 'concurrent/utility/engine'
-require 'concurrent/synchronization/abstract_object'
-require 'concurrent/native_extensions' # JavaObject
-require 'concurrent/synchronization/mutex_object'
-require 'concurrent/synchronization/monitor_object'
-require 'concurrent/synchronization/rbx_object'
-require 'concurrent/synchronization/object'
-
-module Concurrent
-  # {include:file:doc/synchronization.md}
-  module Synchronization
-  end
-end
-
